move things around a bit per comments
